feat: update clueless-interactions to version 1.0.13 and refactor activity handling in API routes
this commit refactors activities to use the interactions tracker library. Instead of updating interactions as you go, they are generated dynamically and then saved as completed whenever the day is over. this makes it so that you can go back and change things like the max time between inputs and have the activity actually be updated properly