feat: implement company weights, refactor into multiple files
move get recommendations to dedicated get recommendations utils file. also creates getcompany weights which weights companies based on the Users goals. if a user has a company under their goal, then it is more likely to show up in recommended problems. this also means that the fetch for getting interviews from DB was moved to a fetch user which includes goals and interviews